Add files via upload
Lab, using the Azure portal, configure the Cloud Shell and download and run the lab setup script. Next import the resource group. Then add AKS, variables, and outputs to the configuration. Lastly, deploy your Kubernetes cluster resources and verify that the cluster is up and healthy.
This lab is inspired by acloud guru